Data Dictionary
Rudimentarily done in a table, with some things needing to be added or likely changed (looking at Global CPI rise). I also doubt that the table is the correct format but oh well, it can be done faster and better later.

Finally moving on to Producing and Implementing, a bit uneasy with the falling behind due to sickness but it's whatever. I'm good at crunching at home. Symbolically created data module.py